requirements.txt
altair==4.2.2
faiss-cpu==1.7.4
google-cloud-storage==2.14.0
langchain==0.0.352
langchain-community==0.0.6
langchain-google-vertexai==0.0.1.post1
pandas==2.1.4
Pillow==9.5.0
PyMuPDF==1.23.8
PyPDF2==3.0.1
pdf2image==1.17.0
streamlit==1.24.0
streamlit-extras==0.4.0
pycryptodome==3.15.0